requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
